evl/sched: tp: fix handling of input/output setup data
tp_control() did not receive the proper input on tp_install requests,
neither did it return the proper information on tp_get. Great, fix
the mess.

+++ include/evl/sched.h
@@ -163,8 +163,8 @@ struct evl_sched_class {
 	void (*sched_kick)(struct evl_thread *thread);
 	ssize_t (*sched_show)(struct evl_thread *thread,
 			      char *buf, ssize_t count);
-	int (*sched_control)(int cpu, union evl_sched_ctlparam *ctlp,
-			union evl_sched_ctlinfo *infp);
+	ssize_t (*sched_control)(int cpu, union evl_sched_ctlparam *ctlp,
+				union evl_sched_ctlinfo *infp);
 	int nthreads;
 	struct evl_sched_class *next;
 	int weight;
